Trigger to resource
