+++ constants.js
@@ -1,4 +1,4 @@
 module.exports = {
   ApiRoot: 'https://www.dcard.tw/_api',
-  WebRoot: 'https://www.dcard.tw/'
+  WebRoot: 'https://www.dcard.tw'
 };
